Moved `NativeBinaryContainer` out of `o.g.language.swift` as it is no longer Swift specific.
