+++ subprojects/language-native/src/main/java/org/gradle/language/swift/SwiftComponent.java
@@ -24,6 +24,7 @@ import org.gradle.api.file.FileCollection;
 import org.gradle.api.provider.Property;
 import org.gradle.api.provider.Provider;
 import org.gradle.language.ComponentWithBinaries;
+import org.gradle.language.NativeBinaryContainer;
 
 /**
  * Configuration for a Swift component, such as a library or executable, defining the source files that make up the component plus other settings.
